Normalise URL, add utils.py file
